+++ src/components/App.js
@@ -1,31 +1,14 @@
 import logo from "../assets/images/logo.svg";
 import "./App.css";
+import SignIn from "./SignIn";
 import Footer from "./Footer";
 
-import withFirebaseAuth from "react-with-firebase-auth";
-import "firebase/auth";
-import firebase from "../firebase";
-
-const firebaseAppAuth = firebase.auth();
-const providers = {
-  emailProvider: new firebase.auth.EmailAuthProvider(),
-  googleProvider: new firebase.auth.GoogleAuthProvider(),
-};
-
-function App({ user, signOut, signInWithEmail, signInWithGoogle }) {
+function App() {
   return (
     <div className="App">
       <header className="App-header">
         <img src={logo} className="App-logo" alt="logo" />
-        {user ? <p>Hello, {user.displayName}</p> : <p>Please sign in.</p>}
-        {user ? (
-          <button onClick={signOut}>Sign out</button>
-        ) : (
-          <>
-            <button onClick={signInWithEmail}>Sign in with email</button>
-            <button onClick={signInWithGoogle}>Sign in with Google</button>
-          </>
-        )}
+        <SignIn />
         <p>
           Edit <code>src/App.js</code> and save to reload.
         </p>
@@ -43,8 +26,4 @@ function App({ user, signOut, signInWithEmail, signInWithGoogle }) {
   );
 }
 
-//export default App;
-export default withFirebaseAuth({
-  providers,
-  firebaseAppAuth,
-})(App);
+export default App;
